feat: move gotToRow behaviour to RowNavigation module
